+++ src/pipx/commands/reinstall.py
@@ -1,6 +1,6 @@
+import sys
 from pathlib import Path
 from typing import List
-import sys
 
 from pipx.commands.inject import inject
 from pipx.commands.install import install
@@ -10,21 +10,21 @@ from pipx.venv import Venv, VenvContainer
 
 
 def reinstall(
-    *, venv_dir: Path, package: str, local_bin_dir: Path, python: str, verbose: bool,
+    *, venv_dir: Path, local_bin_dir: Path, python: str, verbose: bool,
 ):
     venv = Venv(venv_dir, verbose=verbose)
 
     if venv.pipx_metadata.main_package.package_or_url is not None:
         package_or_url = venv.pipx_metadata.main_package.package_or_url
     else:
-        package_or_url = package
+        package_or_url = venv.main_package_name
 
-    uninstall(venv_dir, package, local_bin_dir, verbose)
+    uninstall(venv_dir, local_bin_dir, verbose)
 
     # install main package first
     install(
         venv_dir,
-        package,
+        venv.main_package_name,
         package_or_url,
         local_bin_dir,
         python,
@@ -33,6 +33,7 @@ def reinstall(
         verbose,
         force=True,
         include_dependencies=venv.pipx_metadata.main_package.include_dependencies,
+        suffix=venv.pipx_metadata.main_package.suffix,
     )
 
     # now install injected packages
@@ -44,7 +45,7 @@ def reinstall(
             # This should never happen, but package_or_url is type
             #   Optional[str] so mypy thinks it could be None
             raise PipxError(
-                f"Internal Error injecting package {injected_package} into {package}"
+                f"Internal Error injecting package {injected_package} into {venv_dir.name}"
             )
         inject(
             venv_dir,
@@ -68,20 +69,18 @@ def reinstall_all(
 ):
     failed: List[str] = []
     for venv_dir in venv_container.iter_venv_dirs():
-        package = venv_dir.name
-        if package in skip:
+        if venv_dir.name in skip:
             continue
         try:
             reinstall(
                 venv_dir=venv_dir,
-                package=package,
                 local_bin_dir=local_bin_dir,
                 python=python,
                 verbose=verbose,
             )
         except PipxError as e:
             print(e, file=sys.stderr)
-            failed.append(package)
+            failed.append(venv_dir.name)
     if len(failed) > 0:
         raise PipxError(
             f"The following package(s) failed to reinstall: {', '.join(failed)}"