+++ tests/test_list.py
@@ -1,4 +1,6 @@
-from helpers import run_pipx_cli
+import pytest  # type: ignore
+
+from helpers import mock_legacy_venv, run_pipx_cli
 from pipx import constants, util
 
 
@@ -27,7 +29,28 @@ def test_missing_interpreter(pipx_temp_env, monkeypatch, capsys):
 def test_list_suffix(pipx_temp_env, monkeypatch, capsys):
     suffix = "_x"
     assert not run_pipx_cli(["install", "pycowsay", f"--suffix={suffix}"])
-    assert not run_pipx_cli(["list"])
 
+    assert not run_pipx_cli(["list"])
+    captured = capsys.readouterr()
+    assert f"package pycowsay 0.0.0.1 (pycowsay{suffix})," in captured.out
+
+
+@pytest.mark.parametrize("metadata_version", [None, "0.1"])
+def test_list_legacy_venv(pipx_temp_env, monkeypatch, capsys, metadata_version):
+    assert not run_pipx_cli(["install", "pycowsay"])
+    mock_legacy_venv("pycowsay", metadata_version=metadata_version)
+
+    assert not run_pipx_cli(["list"])
+    captured = capsys.readouterr()
+    assert "package pycowsay 0.0.0.1," in captured.out
+
+
+@pytest.mark.parametrize("metadata_version", ["0.1"])
+def test_list_suffix_legacy_venv(pipx_temp_env, monkeypatch, capsys, metadata_version):
+    suffix = "_x"
+    assert not run_pipx_cli(["install", "pycowsay", f"--suffix={suffix}"])
+    mock_legacy_venv(f"pycowsay{suffix}", metadata_version=metadata_version)
+
+    assert not run_pipx_cli(["list"])
     captured = capsys.readouterr()
     assert f"package pycowsay 0.0.0.1 (pycowsay{suffix})," in captured.out